ROS Tests refactoring
